search & cards updated
