+++ backend/.env
@@ -1,5 +1,5 @@
 
 MONGO_URI: "mongodb+srv://agrorent-admin:I5F3KVAXCn5pfhB6@agrorent.cubmi8z.mongodb.net/agrorent"
 
-MONGO_URL:"mongodb+srv://gurukudte:@geekyzindagi.2qu3wna.mongodb.net/geekyzindagi?retryWrites=true&w=majority"
+#MONGO_URL:"mongodb+srv://gurukudte:@geekyzindagi.2qu3wna.mongodb.net/geekyzindagi?retryWrites=true&w=majority"
