+++ backend/package.json
@@ -15,6 +15,6 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "mongoose": "^7.0.0",
-    "nodemon": "^2.0.21"
+    "nodemon": "^2.0.22"
   }
 }
