+++ backend/server.js
@@ -3,34 +3,23 @@ const dotenv = require("dotenv");
 const cors = require("cors");
 
 const app = express();
-const PORT = process.env.PORT || 5500;
+const PORT = process.env.PORT || 4500;
 dotenv.config();
 
 const connectDB = require("./config/db");
 connectDB();
 
-
 const showAllProducts = require("./routes/showAllProducts");
-
-
 
 app.use(express.json());
 app.use(cors());
 
+app.use("/showallproducts", showAllProducts);
 
-
-app.use('/showallproducts', showAllProducts);
-
-
-
-
-app.get('/', (req,res)=>{
-    res.send('<h1>Welcome to AgroRent</h1>');
-})
-
-
-app.listen(PORT, () => {
-    console.log(`Server Succesfully running on http://localhost:${PORT}`);
+app.get("/", (req, res) => {
+  res.send("<h1>Welcome to AgroRent</h1>");
 });
 
-
+app.listen(PORT, () => {
+  console.log(`Server Succesfully running on http://localhost:${PORT}`);
+});
