+++ frontend/src/App.js
@@ -2,18 +2,12 @@ import Dashboard from "./pages/dashboard/Dashboard";
 
 import LandingPage from "./pages/LandingPage/LandingPage";
 
-
-
-import Request from "./pages/Request";
 function App() {
   return (
     <div className="App">
-
       {/* <Dashboard  /> */}
 
-      <LandingPage/>
-      {/* <Request /> */}
-
+      <LandingPage />
     </div>
   );
 }
