+++ frontend/src/components/card/ToolCard.css
@@ -1,77 +1,76 @@
-.dashboard-page-content-main-content-cards{
-    z-index: -1;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    min-height: 20rem;
-    max-height: 25rem;
+.dashboard-page-content-main-content-cards {
+  z-index: -1;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  min-height: 20rem;
+  max-height: 25rem;
 }
-.dashboard-page-content-main-content-cards-card{
-    width: 22rem;
-    border-radius: 0.3rem;
-    margin: 0.5rem;
-    padding: 0.5rem;
-    background: #EFEFEF;
+.dashboard-page-content-main-content-cards-card {
+  width: 22rem;
+  border-radius: 0.3rem;
+  margin: 0.5rem;
+  padding: 0.5rem;
+  background: #efefef;
 }
-.dashboard-page-content-main-content-cards-card-header-bg{
-    padding: 0 0.5rem;
-    height: 14rem;
-    display: flex;
-    justify-content: flex-start;
-    align-items: flex-end;
-    border-radius: 0.5rem;
-}
-
-.dashboard-page-content-main-content-cards-card-header-img{
-    position:absolute;
-    display: flex;
-    padding: 0.4rem;
-    background: white;
-    width: 6rem;
-    justify-content: center;
-    align-items: center;
-    border-radius: 1rem;
-    z-index: 999;
-}
-.dashboard-page-content-main-content-cards-card-header-img img{
-    height: 5rem;
-    width: 5rem;
-    border-radius: 0.5rem;
-}
-.dashboard-page-content-main-content-cards-card-header-title{
-    padding: 0 1rem;
-    background: white;
-    display: flex;
-    justify-content: flex-end;
-    border-radius: 1rem;
-    align-items: center;
-    margin-left: 4rem;
-    padding-left: 2.3rem;
-}
-.dashboard-page-content-main-content-cards-card-header-title div h4{
-    font-size: 1rem;
-    font-weight: 500;
-}
-.dashboard-page-content-main-content-cards-card-header-title div h5{
-    font-size: 0.8rem;
-    font-weight: 100;
-}
-.dashboard-page-content-main-content-cards-card-header-title h3{
-    font-size: 0.9rem;
-    color: orange;
-    padding-left: 0.5rem;
-}
-.dashboard-page-content-main-content-cards-card-discription{
-    margin-top:0.5rem;
-    padding: 0.5rem;
-}
-.dashboard-page-content-main-content-cards-card-discription h5{
-    font-size:1.2rem;
-    font-weight: 700;
-}
-.dashboard-page-content-main-content-cards-card-discription p{
-    padding-top:0.5rem;
-    font-size:0.9rem;
-    color: grey;
+.dashboard-page-content-main-content-cards-card-header-bg {
+  padding: 0 0.5rem;
+  height: 14rem;
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-end;
+  border-radius: 0.5rem;
 }
 
+.dashboard-page-content-main-content-cards-card-header-img {
+  position: absolute;
+  display: flex;
+  padding: 0.4rem;
+  background: white;
+  width: 6rem;
+  justify-content: center;
+  align-items: center;
+  border-radius: 1rem;
+  z-index: 999;
+}
+.dashboard-page-content-main-content-cards-card-header-img img {
+  height: 5rem;
+  width: 5rem;
+  border-radius: 0.5rem;
+}
+.dashboard-page-content-main-content-cards-card-header-title {
+  padding: 0 1rem;
+  background: white;
+  display: flex;
+  justify-content: flex-end;
+  border-radius: 1rem;
+  align-items: center;
+  margin-left: 4rem;
+  padding-left: 2.3rem;
+}
+.dashboard-page-content-main-content-cards-card-header-title div h4 {
+  font-size: 1rem;
+  font-weight: 500;
+}
+.dashboard-page-content-main-content-cards-card-header-title div h5 {
+  font-size: 0.8rem;
+  font-weight: 100;
+}
+.dashboard-page-content-main-content-cards-card-header-title h3 {
+  font-size: 0.9rem;
+  color: orange;
+  padding-left: 0.5rem;
+}
+.dashboard-page-content-main-content-cards-card-discription {
+  margin-top: 0.5rem;
+  padding: 0.5rem;
+}
+.dashboard-page-content-main-content-cards-card-discription h5 {
+  font-size: 1.2rem;
+  font-weight: 700;
+}
+.dashboard-page-content-main-content-cards-card-discription p {
+  padding-top: 0.5rem;
+  font-size: 0.9rem;
+  color: grey;
+}
