+++ frontend/src/components/card/ToolCard.jsx
@@ -1,38 +1,38 @@
 import React from "react";
 import "./ToolCard.css";
-// import db from "../../db/db.json";
 
 const ToolCard = ({ data, index }) => {
-  console.log(data.toolname);
+  console.log(data.name);
   return (
     <div className="dashboard-page-content-main-content-cards">
       <>
         <div
-          key={data.toolId}
-          className="dashboard-page-content-main-content-cards-card" style={{float:"left"}}
+          key={data.id}
+          className="dashboard-page-content-main-content-cards-card"
+          style={{ float: "left" }}
         >
           <div
             style={{
-              background: `url(${data.bg})`,
+              background: `url(${data.toolbgImg})`,
               backgroundSize: "contain",
               backgroundRepeat: "no-repeat",
             }}
             className="dashboard-page-content-main-content-cards-card-header-bg"
           >
             <div className="dashboard-page-content-main-content-cards-card-header-img">
-              <img src={data.toolimg} alt="" />
+              <img src={data.toolImg} alt="" />
             </div>
             <div className="dashboard-page-content-main-content-cards-card-header-title">
               <div>
-                <h4>{data.toolName}</h4>
-                <h5>Category:{data.toolCategory}</h5>
+                <h4>{data.name}</h4>
+                <h5>Category:{data.category}</h5>
               </div>
-              <h3>{data.toolPrice}</h3>
+              <h3>{data.price}</h3>
             </div>
           </div>
           <div className="dashboard-page-content-main-content-cards-card-discription">
             <h5>Discription</h5>
-            <p>{data.toolDesc}</p>
+            <p>{data.description}</p>
           </div>
         </div>
       </>
