+++ frontend/src/components/header/header.css
@@ -66,3 +66,4 @@ Header {
   #input_search {
     width: 21.5rem;
   }
+
