--- frontend/src/pages/Request.css
@@ -1,121 +1,0 @@
-/* .product-card {
-  z-index: 1;
-  position: absolute;
-  width: 900px;
-  height: 600px; */
-  /* margin-top: 150px;
-  margin-left: 439px;
-  margin-bottom: 20px;
-  
-  margin-bottom: 20px;
-  background: #ffffff;
-  border-radius: 25px; */
-  /* border: 0.5px solid gray;
-  left: 439px;
-top: 272px;
-}
-
-.heading-content{
-  position: absolute;
-width: 275px;
-height: 71px;
-left: 425px;
-top: 167px;
-
-} */
-
-/* .product-card {
-    z-index: ;
-  position: absolute;
-  width: 900px;
-  height: 600px;
-  left: 439px;
-  top: 272px;
-
-  background: #ffffff;
-  border-radius: 25px;
-} */
-/* .color1 {
-  position: absolute;
-  width: 1664.68px;
-  height: 811.81px;
-  left: 0px;
-  top: 110px;
-
-  background: rgba(0, 240, 255, 0.15);
-  filter: blur(32px);
-}
-.color2 {
-  position: absolute;
-  width: 728.38px;
-  height: 704.54px;
-  left: 468.81px;
-  top: 728.46px;
-
-  background: rgba(97, 255, 0, 0.15);
-  filter: blur(32px);
-}
-color3 {
-  position: absolute;
-  width: 605.22px;
-  height: 505.89px;
-  left: 953.51px;
-  top: 573.51px;
-
-  background: rgba(255, 66, 66, 0.15);
-  filter: blur(32px);
-}
-
-.gradiants-background {
-  z-index: -1;
-}
-
-.gradinat1 {
-  position: absolute;
-  width: 1664.68px;
-  height: 1323px;
-  left: 0px;
-  top: 110px;
-}
-
-.gradinat2 {
-  position: absolute;
-  width: 1257px;
-  height: 999px;
-  left: -499px;
-  top: -269px;
-}
-.gradinat3 {
-  position: absolute;
-  width: 1257px;
-  height: 999px;
-  left: -499px;
-  top: -382px;
-} */
-
-.product-card {
-  z-index: 1;
-  position: absolute;
-  width: 56.25em;
-  height: 37.5em;
-  /* margin-top: 9.375em;
-  margin-left: 21.859em;
-  margin-bottom: 1.25em;
-  
-  margin-bottom: 1.25em;
-  background: #ffffff;
- */
-
-  border-radius: 1.5625em;
-  border: 0.03125em solid gray;
-  left: 21.859em;
-  top: 16.95em;
-  }
-  
-  .heading-content{
-  position: absolute;
-  width: 17.1875em;
-  height: 4.4375em;
-  left: 21.016em;
-  top: 10.5475em;
-  }
